Split file insert-cf-params
